requirements.txt
redis==2.10
ansicolors>=1.1
scikit-learn==0.18.1
unicodecsv>=0.14